Fixes for x86_64 windows builds.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@1062 e19a5227-2556-0410-816f-a35683127991

+++ apps/zxtune-qt/ui/utils.h
@@ -26,7 +26,7 @@ inline QString ToQString(const String& str)
 #ifdef UNICODE
   return QString::fromStdWString(str);
 #else
-  return QString::fromLocal8Bit(str.data(), str.size());
+  return QString::fromLocal8Bit(str.data(), static_cast<int>(str.size()));
 #endif
 }
